add feature to load/store cache for faster runs

+++ scrumTools/src/main/java/ch/paru/scrumTools/exchangeServer/utils/interceptors/command/CacheFileHandler.java
@@ -15,12 +15,18 @@ import ch.paru.scrumTools.exchangeServer.services.contact.ServerContact;
 import ch.paru.scrumTools.exchangeServer.services.contact.ServerContactGroup;
 import ch.paru.scrumTools.exchangeServer.utils.exceptions.ServerException;
 
+import com.google.common.collect.Maps;
 import com.thoughtworks.xstream.XStream;
 import com.thoughtworks.xstream.io.xml.DomDriver;
 
 public class CacheFileHandler {
 
-	public CacheFileHandler() {
+	private String path;
+	private String cacheName;
+
+	public CacheFileHandler(String path, String cacheName) {
+		this.path = path;
+		this.cacheName = cacheName;
 	}
 
 	private XStream getXStream() {
@@ -35,10 +41,14 @@ public class CacheFileHandler {
 		return xStream;
 	}
 
-	public void store(Map<String, Object> data, String path, String cacheName) {
+	public void store(Map<String, Object> data) {
+		if (!fileExists()) {
+			return;
+		}
+
 		try {
 			XStream xStream = getXStream();
-			File file = new File(getFilePath(path, cacheName));
+			File file = new File(getFilePath());
 			FileWriter writer = new FileWriter(file, false);
 			ObjectOutputStream out = xStream.createObjectOutputStream(writer);
 			out.writeObject(data);
@@ -50,10 +60,14 @@ public class CacheFileHandler {
 	}
 
 	@SuppressWarnings("unchecked")
-	public Map<String, Object> load(String path, String cacheName) {
+	public Map<String, Object> load() {
+		if (!fileExists()) {
+			return Maps.newHashMap();
+		}
+
 		try {
 			XStream xStream = getXStream();
-			File file = new File(getFilePath(path, cacheName));
+			File file = new File(getFilePath());
 			FileReader reader = new FileReader(file);
 			ObjectInputStream in = xStream.createObjectInputStream(reader);
 			Map<String, Object> data = (Map<String, Object>) in.readObject();
@@ -65,7 +79,12 @@ public class CacheFileHandler {
 		}
 	}
 
-	private String getFilePath(String path, String cacheName) {
+	private String getFilePath() {
 		return path + "cache-" + cacheName + ".xml";
+	}
+
+	private boolean fileExists() {
+		File file = new File(getFilePath());
+		return file.exists();
 	}
 }
